Add package x/feed.
This package provides a simple fixed-capacity buffered item feed.
Writes to the feed are non-blocking up to the capacity.

This is an essay toward a reworking of the Tendermint event system into two
layers: This package is meant to handle the "inner" layer, which gathers and
sequences the raw event records from various parts of the system.

The consumer of the raw event stream will index, filter, buffer, and dispatch
events to other consumers, e.g., RPC clients and pubsub subscribers.